Added extra comments and Documentation
REID_Keras file also got a documentation(not a very good one for now) and also explained the final result which was an overfitting and the reason for it.

Camera got a new Documentation and the reason as to why it is not functional at the moment.

Formatting in i_LIDS_VID has been corrected